requirements.txt
appdirs==1.4.4
asgiref==3.2.7
Authlib==0.14.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
distlib==0.3.0
Django==3.0.6
django-cors-middleware==1.5.0
django-filter==2.2.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
Markdown==3.2.2
oauthlib==3.1.0
pbr==5.4.5
pycparser==2.20
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
